Timing: Make `LoopingTimeSource` double based

+++ Furball.Engine/Engine/Timing/LoopingTimeSource.cs
@@ -2,9 +2,9 @@ namespace Furball.Engine.Engine.Timing;
 
 public class LoopingTimeSource : ITimeSource {
     public ITimeSource BaseTimeSource;
-    public int         LoopTime;
+    public double      LoopTime;
         
-    public LoopingTimeSource(ITimeSource baseTimeSource, int loopTime) {
+    public LoopingTimeSource(ITimeSource baseTimeSource, double loopTime) {
         this.BaseTimeSource = baseTimeSource;
         this.LoopTime       = loopTime;
     }
